requirements.txt
chainer==7.8.0
hydra-core==1.1.1
ipython==8.2.0
lmdb==1.3.0
msgpack_numpy==0.4.7.1
numpy==1.21.2
omegaconf==2.1.1
open3d==0.13.0
pymeshlab==2021.10
pytorch_lightning==1.6.0
setuptools==58.0.4
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
